+++ R/CreateArgFunctions.R
@@ -337,7 +337,7 @@ createStratifyByPsAndCovariatesArgs <- function(numberOfStrata = 5,
 #' @param includeCovariateIds       Include only these covariates in the outcome model.
 #' @param profileGrid               A one-dimensional grid of points on the log(relative risk) scale
 #'                                  where the likelihood for the treatment variable coefficient is
-#'                                  sampled.
+#'                                  sampled. Set to NULL to skip profiling.
 #' @param prior                     The prior used to fit the model. See Cyclops::createPrior() for
 #'                                  details.
 #' @param control                   The control object used to control the cross-validation used to